+++ Source/Lutefisk3D/2D/CollisionChain2D.h
@@ -53,7 +53,7 @@ public:
     /// Return loop.
     bool GetLoop() const { return loop_; }
     /// Return vertex count.
-    unsigned GetVertexCount() const { return vertices_.size(); }
+    size_t GetVertexCount() const { return vertices_.size(); }
     /// Return vertex.
     const Vector2& GetVertex(unsigned index) const { return (index < vertices_.size()) ? vertices_[index] : Vector2::ZERO; }
     /// Return vertices.
