+++ Source/Lutefisk3D/Graphics/DecalSet.h
@@ -124,7 +124,7 @@ public:
     /// Return material.
     Material* GetMaterial() const;
     /// Return number of decals.
-    unsigned GetNumDecals() const { return decals_.size(); }
+    size_t GetNumDecals() const { return decals_.size(); }
     /// Retur number of vertices in the decals.
     unsigned GetNumVertices() const { return numVertices_; }
     /// Retur number of vertex indices in the decals.
